对订单详情更改，准备更改乘车界面的需求变更，如果乘车界面的需求又回退，则可回退到该版本

+++ EBus/Classes/Me/Controller/FreeCertificate/EBFreeCertificateController.m
@@ -409,7 +409,7 @@
         label.backgroundColor = EB_RGBColor(213, 213, 213);
         label.textColor = [UIColor whiteColor];
         label.textAlignment = NSTextAlignmentCenter;
-        label.text = [NSString stringWithFormat:@"%d",i + 1];
+        label.text = [NSString stringWithFormat:@"%ld",(long)(i + 1)];
         [self.firView addSubview:label];
         [self.firLabels addObject:label];
     }
